move test script to Samples
